pyproject.toml
[project]
name = "purchase-predict-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.3",
    "google-cloud-storage>=3.10.1",
    "gunicorn>=23.0.0",
    "lightgbm>=4.6.0",
    "mlflow==3.0.1",
    "pandas<3.0.0",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
]
